+++ cmn-data-svc-server/src/main/java/org/acumos/cds/service/AbstractSearchServiceImpl.java
@@ -37,7 +37,6 @@ import org.springframework.data.domain.Sort;
  */
 public abstract class AbstractSearchServiceImpl {
 
-
 	@Autowired
 	protected EntityManager entityManager;